Refactor the view license page
All data to be retrieved per tab item to speed up page load time

Add active tab management in the service to handle complex logic

Build up tabs in service to allow extensions and control roll based access at the URl level (Not in the render as currently done)
