+++ app/views/licences/tab-group.njk
@@ -1,0 +1,52 @@
+<div class="govuk-tabs">
+    <h2 class="govuk-tabs__title">
+        Contents
+    </h2>
+    <ul class="govuk-tabs__list">
+        {% for tab in tabs %}
+            <li class="govuk-tabs__list-item {% if tab.active === true %} govuk-tabs__list-item--selected {% endif %}">
+                <a class="govuk-tabs__tab" href="{{ tab.href }}">
+                    {{ tab.name }}
+                </a>
+            </li>
+        {% endfor %}
+
+        {#        <li class="govuk-tabs__list-item govuk-tabs__list-item--selected"> #}
+        {#            <a class="govuk-tabs__tab" href="summary"> #}
+        {#                Summary #}
+        {#            </a> #}
+        {#        </li> #}
+        {#        <li class="govuk-tabs__list-item"> #}
+        {#            <a class="govuk-tabs__tab" href="contact-details"> #}
+        {#                Contact details #}
+        {#            </a> #}
+        {#        </li> #}
+        {#        <li class="govuk-tabs__list-item"> #}
+        {#            <a class="govuk-tabs__tab" href="returns"> #}
+        {#                Returns #}
+        {#            </a> #}
+        {#        </li> #}
+        {#        <li class="govuk-tabs__list-item"> #}
+        {#            <a class="govuk-tabs__tab" href="communications"> #}
+        {#                Communications #}
+        {#            </a> #}
+        {#        </li> #}
+        {#        <li class="govuk-tabs__list-item"> #}
+        {#            <a class="govuk-tabs__tab" href="bills"> #}
+        {#                Bills #}
+        {#            </a> #}
+        {#        </li> #}
+        {#        <li class="govuk-tabs__list-item"> #}
+        {#            <a class="govuk-tabs__tab" href="charge-information"> #}
+        {#                Charge information #}
+        {#            </a> #}
+        {#        </li> #}
+
+    </ul>
+
+    <div class="govuk-tabs__panel" id="{{ activeTab }}">
+        {% if activeTab === 'summary' %}
+            {% include "licences/tabs/summary.njk" %}
+        {% endif %}
+    </div>
+</div>
